@@ -33,13 +33,13 @@
     "@mui/icons-material": "^5.15.1",
     "@mui/material": "^5.15.1",
     "@mui/system": "^5.15.1",
-    "@mui/x-data-grid": "^6.9.0",
+    "@mui/x-data-grid": "^6.18.5",
     "@olympusdao/component-library": "3.1.7",
     "@olympusdao/treasury-subgraph-client": "^1.1.0",
     "@rainbow-me/rainbowkit": "^0.12.18",
     "@reduxjs/toolkit": "^1.9.3",
     "@tanstack/react-query": "^4.2.3",
-    "@types/recharts": "^1.8.24",
+    "@types/recharts": "^1.8.29",
     "@vitest/coverage-v8": "^0.34.6",
     "@wundergraph/react-query": "^0.9.27",
     "axios": "^1.6.2",
@@ -49,16 +49,16 @@
     "get-value": "^3.0.1",
     "graphql-request": "^6.1.0",
     "luxon": "^3.4.4",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-ga": "^3.3.1",
     "react-ga4": "^2.1.0",
-    "react-hot-toast": "^2.4.0",
+    "react-hot-toast": "^2.4.1",
     "react-redux": "^8.0.5",
     "react-router-dom": "^6.21.1",
     "recharts": "^2.10.3",
     "tinycolor2": "^1.6.0",
-    "typescript": "^5.2.2",
+    "typescript": "^5.3.3",
     "wagmi": "^0.12.19"
   },
   "devDependencies": {
@@ -68,14 +68,14 @@
     "@redocly/cli": "^1.6.0",
     "@tanstack/react-query-devtools": "^4.19.1",
     "@testing-library/dom": "^9.3.3",
-    "@testing-library/react": "^12.1.5",
-    "@typechain/ethers-v5": "^10.2.1",
-    "@types/css-mediaquery": "^0.1.1",
+    "@testing-library/react": "^14.1.2",
+    "@typechain/ethers-v5": "^11.1.2",
+    "@types/css-mediaquery": "^0.1.4",
     "@types/get-value": "^3.0.5",
     "@types/luxon": "^3.3.7",
     "@types/node": "^20.10.5",
-    "@types/react": "^18.0.28",
-    "@types/react-dom": "^18.0.3",
+    "@types/react": "^18.2.45",
+    "@types/react-dom": "^18.2.18",
     "@types/react-router-dom": "^5.3.3",
     "@types/tinycolor2": "^1.4.6",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
@@ -84,19 +84,19 @@
     "@vitest/ui": "^0.34.7",
     "css-mediaquery": "^0.1.2",
     "dotenv": "^16.3.1",
-    "eslint": "^8.43.0",
+    "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-no-relative-import-paths": "^1.5.3",
-    "eslint-plugin-prettier": "^5.0.0",
+    "eslint-plugin-prettier": "^5.1.1",
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-simple-import-sort": "^10.0.0",
-    "eslint-plugin-unused-imports": "^2.0.0",
+    "eslint-plugin-unused-imports": "^3.0.0",
     "fast-check": "^3.15.0",
     "happy-dom": "^12.10.3",
     "husky": "^8.0.3",
     "jsdom": "^21.1.1",
-    "lint-staged": "^14.0.1",
+    "lint-staged": "^15.2.0",
     "orval": "^6.23.0",
     "prettier": "^3.1.1",
     "prop-types": "^15.8.1",
