@@ -1,3 +1,7 @@
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import Root from "src/Root";
-ReactDOM.render(<Root />, document.getElementById("root"));
+
+const container = document.getElementById("root");
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+const root = createRoot(container!);
+root.render(<Root />);
